Code cleanup based on code-review
