+++ opensrp-app/src/main/java/org/smartregister/login/interactor/BaseLoginInteractor.java
@@ -10,6 +10,7 @@ import org.json.JSONException;
 import org.json.JSONObject;
 import org.smartregister.AllConstants;
 import org.smartregister.CoreLibrary;
+import org.smartregister.P2POptions;
 import org.smartregister.R;
 import org.smartregister.domain.LoginResponse;
 import org.smartregister.domain.TimeStatus;
@@ -221,8 +222,11 @@ public abstract class BaseLoginInteractor implements BaseLoginContract.Interacto
      * Call super if you override this method.
      */
     protected void scheduleJobsImmediately() {
-        // Finish processing any unprocessed sync records here
-        P2pServiceJob.scheduleJobImmediately(P2pServiceJob.TAG);
+        P2POptions p2POptions = CoreLibrary.getInstance().getP2POptions();
+        if (p2POptions != null && p2POptions.isEnableP2PLibrary()) {
+            // Finish processing any unprocessed sync records here
+            P2pServiceJob.scheduleJobImmediately(P2pServiceJob.TAG);
+        }
 
         if (NetworkUtils.isNetworkAvailable()) {
             PullUniqueIdsServiceJob.scheduleJobImmediately(PullUniqueIdsServiceJob.TAG);
