+++ sample/src/main/java/org/smartregister/sample/MainActivity.java
@@ -55,9 +55,4 @@ public class MainActivity extends AppCompatActivity {
 
         return super.onOptionsItemSelected(item);
     }
-
-    @Override
-    protected void onResume() {
-        super.onResume();
-    }
 }
